// frontend/src/api/models/albumDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ArtistSummaryDto } from './artistSummaryDto';
import type { TrackListDto } from './trackListDto';

export interface AlbumDto {
  id?: string;
  title?: string;
  coverImageUrl?: string;
  releaseDate?: string;
  spotifyId?: string;
  artist?: ArtistSummaryDto;
  trackCount?: number;
  tracks?: TrackListDto[];
}